Fix scroll locking in ShadowDOM
Scroll locking was broken in ShadowDOM, because the result of calling
`element.contains` is different compared to ShadeyDOM (which lacks
encapsulation). What we want in the scroll manager is to query the
composed ShadowDOM tree when testing element hierarchy. So, a new method
and supporting caching mechanism has been added to enable this in a
(hopefully) performant manner.